+++ client/src/components/select/index.tsx
@@ -6,15 +6,20 @@ interface Props {
   onChange: any;
   options: SelectProps["options"];
   defaultValue?: SelectProps["defaultValue"];
+  value?: SelectProps["value"];
+  onFocus?: SelectProps["onFocus"];
+  isMulti?: boolean;
 }
 
-const Select: React.FC<Props> = ({ onChange, options, defaultValue }) => {
+const Select: React.FC<Props> = ({ onChange, onFocus, isMulti, options, defaultValue, value }) => {
   const [menuOpen, setMenuOpen] = React.useState(false);
 
   return (
     <ReactSelect
+      onFocus={onFocus}
+      value={value}
       isSearchable
-      isMulti
+      isMulti={isMulti !== undefined ? isMulti : true}
       styles={SelectStyles}
       onChange={onChange}
       options={options}